# api-docs.json
{"openapi":"3.0.1","info":{"title":"FlavorDB 2.0 API","description":"API for accessing molecule and flavor information","contact":{"name":"FlavorDB Team","email":"contact@flavordb.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/properties/taste-threshold":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByTasteThreshold","parameters":[{"name":"values","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/synthesis":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesBySynthesis","parameters":[{"name":"values","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-tradeAssociationGuidelines":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByTradeAssociationGuidelines","parameters":[{"name":"guideline","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-naturalOccurrence":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByNaturalOccurrence","parameters":[{"name":"occurrence","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-nas":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByNasNo","parameters":[{"name":"nas_no","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-jecfa":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByJecfaNo","parameters":[{"name":"jecfa_no","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-iofi-categorisation":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByIofiCategorisation","parameters":[{"name":"category","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-flNo":{"get":{"tags":["fn-property-controller"],"operationId":"getFnPropertiesByFlNo","parameters":[{"name":"fl_no","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-fema":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByFemaNo","parameters":[{"name":"fema_no","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-einecs":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByEinecsNo","parameters":[{"name":"einecs_no","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-description":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByDescription","parameters":[{"name":"desc","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-coe":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByCoeNo","parameters":[{"name":"coe_no","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-coe-approval":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByCoeApproval","parameters":[{"name":"coe_approval","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/properties/by-aromaThresholdValues":{"get":{"tags":["fn-property-controller"],"operationId":"getPropertiesByAromaThresholdValues","parameters":[{"name":"threshold","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FnProperty"}}}}}}}},"/api/more_properties/by-surfaceArea":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesBySurfaceArea","parameters":[{"name":"surface_area","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-rotatableBonds":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByNumRotatableBonds","parameters":[{"name":"rotatable_bonds","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-numberOfAtoms":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByNumberOfAtoms","parameters":[{"name":"number_of_atoms","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-numRings":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByNumRings","parameters":[{"name":"num_rings","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-energy":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByEnergy","parameters":[{"name":"energy","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-aromaticRings":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByNumberOfAromaticRings","parameters":[{"name":"aromatic_rings","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-aromaticBonds":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByNumberOfAromaticBonds","parameters":[{"name":"aromatic_bonds","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/more_properties/by-alogp":{"get":{"tags":["more-property-controller"],"operationId":"getPropertiesByAlogp","parameters":[{"name":"alogp","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoreProperty"}}}}}}}},"/api/molecules_data/filter-by-weight-range":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByWeightRange","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"to","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/filter-by-weight-from":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByWeightFrom","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/filter-by-type":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByType","parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/filter-by-hbd-count":{"get":{"tags":["molecule-controller"],"operationId":"filterByHbdCount","parameters":[{"name":"hbdCount","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/filter-by-hba-count":{"get":{"tags":["molecule-controller"],"operationId":"filterByHbaCount","parameters":[{"name":"hbaCount","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-topologicalPolarSurfaceArea":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByTopologicalPolarSurfaceArea","parameters":[{"name":"topological_polar_surface_area","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-pubchemId":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByPubchemId","parameters":[{"name":"pubchem_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-monoisotopicMass":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByMonoisotopicMass","parameters":[{"name":"monoisotopic_mass","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-heavyAtomCount":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByHeavyAtomCount","parameters":[{"name":"heavy_atom_count","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-functionalGroups":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByFunctionalGroups","parameters":[{"name":"functional_groups","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-flavorProfile":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByFlavorProfile","parameters":[{"name":"flavor_profile","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-femaFlavorProfile":{"get":{"tags":["molecule-controller"],"operationId":"getMoleculesByFemaFlavorProfile","parameters":[{"name":"fema_flavor_profile","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/molecules_data/by-commonName":{"get":{"tags":["molecule-controller"],"operationId":"getMolecules","parameters":[{"name":"common_name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Molecule"}}}}}}}},"/api/entities/by-natural-source":{"get":{"tags":["entity-controller"],"operationId":"getEntitiesByNaturalSourceName","parameters":[{"name":"natural_source_name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Entity"}}}}}}}},"/api/entities/by-name-and-category":{"get":{"tags":["entity-controller"],"operationId":"getEntitiesByNameAndCategory","parameters":[{"name":"entity_alias","in":"query","required":true,"schema":{"type":"string"}},{"name":"category","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Entity"}}}}}}}},"/api/entities/by-entity-alias-readable":{"get":{"tags":["entity-controller"],"operationId":"getEntitiesByEntityAliasReadable","parameters":[{"name":"entity_alias_readable","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Entity"}}}}}}}}},"components":{"schemas":{"FnProperty":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"botanicalName":{"type":"string"},"botanicalFamily":{"type":"string"},"synonyms":{"type":"string"},"casNo":{"type":"string"},"flNo":{"type":"string"},"femaNo":{"type":"number","format":"double"},"nasNo":{"type":"string"},"coeNo":{"type":"string"},"coeApproval":{"type":"string"},"einecsNo":{"type":"string"},"fda":{"type":"string"},"femaGras":{"type":"string"},"jecfaNo":{"type":"string"},"jecfaStatus":{"type":"string"},"additives":{"type":"string"},"ifraCategory":{"type":"string"},"reachNo":{"type":"string"},"scentRefLibrary":{"type":"string"},"fct":{"type":"string"},"notes":{"type":"string"},"derivatives":{"type":"string"},"derivativeNames":{"type":"string"},"note":{"type":"string"},"foreignNames":{"type":"string"},"essentialOilComposition":{"type":"string"},"genusSpecies":{"type":"string"},"composition":{"type":"string"},"insNo":{"type":"string"},"animalFamily":{"type":"string"},"physicalChemicalCharacteristics":{"type":"string"},"naturalOccurrence":{"type":"string"},"tradeAssociationGuidelines":{"type":"string"},"aromaThresholdValues":{"type":"string"},"description":{"type":"string"},"iofiCategorisation":{"type":"string"}}},"MoreProperty":{"type":"object","properties":{"id":{"type":"string"},"pubchemId":{"type":"integer","format":"int32"},"numberOfAtoms":{"type":"integer","format":"int32"},"molecularFormula":{"type":"string"},"molecularComposition":{"type":"string"},"molecularWeight":{"type":"number","format":"double"},"molecularMass":{"type":"number","format":"double"},"energy":{"type":"number","format":"double"},"alogp":{"type":"number","format":"double"},"logd":{"type":"number","format":"double"},"molecularSolubility":{"type":"number","format":"double"},"pka":{"type":"string"},"numberOfAromaticBonds":{"type":"integer","format":"int32"},"numberOfAromaticRings":{"type":"integer","format":"int32"},"hbaCount":{"type":"integer","format":"int32"},"hbdCount":{"type":"integer","format":"int32"},"numberOfHAcceptor":{"type":"integer","format":"int32"},"numberOfHAcceptorLipinski":{"type":"integer","format":"int32"},"numberOfHDonor":{"type":"integer","format":"int32"},"numberOfHDonorLipinski":{"type":"integer","format":"int32"},"numberOfHBonds":{"type":"integer","format":"int32"},"numRings":{"type":"integer","format":"int32"},"numRotatableBonds":{"type":"integer","format":"int32"},"surfaceArea":{"type":"number","format":"double"},"molecularSasa":{"type":"number","format":"double"},"radiusOfGyration":{"type":"number","format":"double"},"molecular3dSasa":{"type":"number","format":"double"}}},"Molecule":{"type":"object","properties":{"id":{"type":"string"},"pubchemId":{"type":"integer","format":"int32"},"iupacName":{"type":"string"},"commonName":{"type":"string"},"smile":{"type":"string"},"molecularWeight":{"type":"number","format":"double"},"hbdCount":{"type":"integer","format":"int32"},"hbaCount":{"type":"integer","format":"int32"},"numRotatableBonds":{"type":"integer","format":"int32"},"complexity":{"type":"number","format":"double"},"topologicalPolarSurfaceArea":{"type":"number","format":"double"},"monoisotopicMass":{"type":"number","format":"double"},"exactMass":{"type":"number","format":"double"},"xlogp":{"type":"number","format":"double"},"charge":{"type":"integer","format":"int32"},"heavyAtomCount":{"type":"integer","format":"int32"},"atomStereoCount":{"type":"integer","format":"int32"},"definedAtomStereoCenterCount":{"type":"integer","format":"int32"},"undefinedAtomStereoCenterCount":{"type":"integer","format":"int32"},"bondStereoCount":{"type":"integer","format":"int32"},"definedBondStereoCenterCount":{"type":"integer","format":"int32"},"undefinedBondStereoCenterCount":{"type":"integer","format":"int32"},"isotopeAtomCount":{"type":"integer","format":"int32"},"covalentlyBondedUnitCount":{"type":"integer","format":"int32"},"casId":{"type":"string"},"femaNumber":{"type":"string"},"femaFlavorProfile":{"type":"string"},"odor":{"type":"string"},"taste":{"type":"string"},"functionalGroups":{"type":"string"},"inchi":{"type":"string"},"volume3d":{"type":"number","format":"double"},"foodDbFlavorProfile":{"type":"string"},"superSweet":{"type":"string"},"bitter":{"type":"number","format":"double"},"superSweetDbId":{"type":"number","format":"double"},"bitterDbId":{"type":"number","format":"double"},"foodDbId":{"type":"string"},"flavorNetId":{"type":"integer","format":"int32"},"fenoroliAndOs":{"type":"integer","format":"int32"},"natural":{"type":"integer","format":"int32"},"unknownNatural":{"type":"integer","format":"int32"},"synthetic":{"type":"integer","format":"int32"},"flavorProfile":{"type":"string"}}},"Entity":{"type":"object","properties":{"id":{"type":"string"},"entityId":{"type":"integer","format":"int32"},"category":{"type":"string"},"categoryReadable":{"type":"string"},"entityAlias":{"type":"string"},"entityAliasBasket":{"type":"string"},"entityAliasReadable":{"type":"string"},"entityAliasSynonyms":{"type":"string"},"entityAliasUrl":{"type":"string"},"naturalSourceName":{"type":"string"},"naturalSourceUrl":{"type":"string"}}}}}}